SuperLUDIST

Documentation for SuperLUDIST.

SuperLUDIST.DistributedSuperMatrix

SuperLUDIST.DistributedSuperMatrix

SuperLUDIST.DistributedSuperMatrix

SuperLUDIST.ReplicatedSuperMatrix

MatrixMarket.mmread

MatrixMarket.mmread

SparseBase.Communication.scatterstore!

SuperLUDIST.DistributedSuperMatrix — Type

mutable struct DistributedSuperMatrix{T, I, S, F, J} <: SuperLUDIST.AbstractSuperMatrix{T, I, S}

SuperMatrix distributed by row between all ranks in a communicator. The Julia representation may be accessed by A.store::SparseBase.CSRStore{T, CIndex{<:Int}}.

Each local matrix stores a subset of rows specified by A.globalrows

Extended

supermatrix

Internal: Reference to Common.SuperMatrix{I}

format

Internal: Reference to storage format Common.NRFormat_loc{I} held by supermatrix

store

Julia-side storage, of type SparseBase.CSRStore. Keeps memory alive.

globalsize

first_row

SuperLUDIST.DistributedSuperMatrix — Method

DistributedSuperMatrix(A::SparseBase.SinglyCompressedStore{<:Any, StorageOrders.RowMajor(), CIndices.CIndex{Ti}}, firstrow, globalsize) -> DistributedSuperMatrix

Construct a DistributedSuperMatrix from a finished SparseBase.CSRStore.

Arguments

A : sparse storage in CSR form, valid index types are {Int32, Int64}, valid element types are {Float32, Float64, ComplexF64}.

firstrow : the 1-based starting row of A on this rank.

globalsize : the size of A across all ranks.

SuperLUDIST.DistributedSuperMatrix — Method

DistributedSuperMatrix{Tv, Ti}()

Construct an empty DistributedSuperMatrix to be filled later.

Note: Currently, modifying Julia-level fields will not propagate to the C structs. For instance in order to update the size you must update the size of A::DistributedSuperMatrix.store as well as the sizes stored in A.format and A.supermatrix.

SuperLUDIST.ReplicatedSuperMatrix — Type

mutable struct ReplicatedSuperMatrix{T, I, S, F, J} <: SuperLUDIST.AbstractSuperMatrix{T, I, S}

SuperMatrix that has been globally replicated on all ranks in a communicator. The local storage contains the full matrix, the Julia representation may be accessed by A.store::SparseBase.CSCStore{T, CIndex{<:Int}}.

Extended

supermatrix

Internal: Reference to Common.SuperMatrix{I}

format

Internal: Reference to storage format Common.NCformat{I} held by supermatrix

store

Julia-side storage, of type SparseBase.CSCStore. Keeps memory alive.

MatrixMarket.mmread — Method

mmread(DistributedSuperMatrix{Tv, Ti}, filename::String, partitioner = distribute_evenly)

Read an mtx file into a compressed sparse column format and broadcast it to all ranks in comm.

To read an mtx file locally use mmread(SparseBase.CSCStore...).

Arguments

arg1<:Type{ReplicatedSuperMatrix{Tv, Ti}}

Tv <: Union{Float32, Float64, ComplexF64}: Non-optional eltype.

Ti <: Union{Int32, Int64}: Non-optional index type.

filename::String: .mtx file accessible by the root process.

Keyword Arguments

desymmetrize::Bool = true: If the matrix file is represented

in symmetric form, represent in the full form.

root::Int = 0: The MPI process on which to read the file.

comm::MPI.Comm = MPI.COMM_WORLD: The MPI communicator on which to distribute

the matrix.

Example

using MPI, SuperLUDIST, MatrixMarket
A = MatrixMarket.mmread(ReplicatedSuperMatrix{Float64, Int32}, )

MatrixMarket.mmread — Method

mmread(ReplicatedSuperMatrix{Tv, Ti}, filename::String)

Read an mtx file into a compressed sparse column format and broadcast it to all ranks in comm.

To read an mtx file locally use mmread(SparseBase.CSCStore...).

Arguments

arg1<:Type{ReplicatedSuperMatrix{Tv, Ti}}

Tv <: Union{Float32, Float64, ComplexF64}: Non-optional eltype.

Ti <: Union{Int32, Int64}: Non-optional index type.

filename::String: .mtx file accessible by the root process.

Keyword Arguments

desymmetrize::Bool = true: If the matrix file is represented

in symmetric form, represent in the full form.

root::Int = 0: The MPI process on which to read the file.

comm::MPI.Comm = MPI.COMM_WORLD: The MPI communicator on which to distribute

the matrix.

Example

using MPI, SuperLUDIST, MatrixMarket
A = MatrixMarket.mmread(ReplicatedSuperMatrix{Float64, Int32}, )

SparseBase.Communication.scatterstore! — Method

scatterstore!(rstore::DistributedSuperMatrix{Tv, Ti}, sstore::Union{Nothing, SparseBase.SinglyCompressedStore{Tvalues, StorageOrders.RowMajor()} where Tvalues}, chunksizes) -> DistributedSuperMatrix