SuperLUDIST

Documentation for SuperLUDIST.

SuperLUDIST.DistributedSuperMatrix

SuperLUDIST.DistributedSuperMatrix

SuperLUDIST.DistributedSuperMatrix

SuperLUDIST.DistributedSuperMatrix

SuperLUDIST.DistributedSuperMatrix

SuperLUDIST.ReplicatedSuperMatrix

MatrixMarket.mmread

MatrixMarket.mmread

SparseBase.Communication.scatterstore!

SuperLUDIST.pgssvx!

SuperLUDIST.pgssvx!

SuperLUDIST.DistributedSuperMatrix — Type

mutable struct DistributedSuperMatrix{T, I, S, F, J, G} <: SuperLUDIST.AbstractSuperMatrix{T, I, S}

SuperMatrix distributed by row between all ranks in a communicator. The Julia representation may be accessed by A.store::SparseBase.CSRStore{T, CIndex{<:Int}}.

Each local matrix stores a subset of rows specified by A.globalrows

Extended

supermatrix

Internal: Reference to Common.SuperMatrix{I}

format

Internal: Reference to storage format Common.NRFormat_loc{I} held by supermatrix

store

Julia-side storage, of type SparseBase.CSRStore. Keeps memory alive.

globalsize

first_row

grid

SuperLUDIST.DistributedSuperMatrix — Method

DistributedSuperMatrix{Tv, Ti}()

Construct an empty DistributedSuperMatrix to be filled later.

Note: Currently, modifying Julia-level fields will not propagate to the C structs. For instance in order to update the size you must update the size of A::DistributedSuperMatrix.store as well as the sizes stored in A.format and A.supermatrix.

SuperLUDIST.DistributedSuperMatrix — Method

DistributedSuperMatrix(rowptr, colidx, v, firstrow, localsize, globalsize, grid::SuperLUDIST.Grid{Ti}) -> DistributedSuperMatrix

Construct a DistributedSuperMatrix from the vectors of a CSR matrix, and the necessary metadata.

Valid index types are {Int32, Int64}, or CIndex{Int32}, CIndex{Int64}, if the indices are already 0-based. Valid element types are {Float32, Float64, ComplexF64}.

Arguments

rowptr, colidx, v : sparse storage vectors, must be valid CSR matrix internals with types noted above.

firstrow : the 1-based starting row of the matrix on this rank.

localsize : the local size of the matrix on this rank.

globalsize : the size of the matrix across all ranks.

grid : the grid on which the matrix is distributed.

SuperLUDIST.DistributedSuperMatrix — Method

DistributedSuperMatrix(A::SparseBase.SinglyCompressedStore{<:Any, StorageOrders.RowMajor(), CIndices.CIndex{Ti}}, firstrow, globalsize, grid::SuperLUDIST.Grid{Ti}) -> DistributedSuperMatrix

Construct a DistributedSuperMatrix from a finished SparseBase.CSRStore.

Arguments

A : sparse storage in CSR form, valid index types are {Int32, Int64}, valid element types are {Float32, Float64, ComplexF64}.

firstrow : the 1-based starting row of A on this rank.

globalsize : the size of A across all ranks.

SuperLUDIST.DistributedSuperMatrix — Method

DistributedSuperMatrix(::Tuple{var"#s19690", var"#s19690"} where var"#s19690"<:(AbstractVector), v, firstrow, localsize, globalsize, grid::SuperLUDIST.Grid{Ti}) -> DistributedSuperMatrix

Construct a DistributedSuperMatrix from COO format and the necessary metadata.

Valid index types are {Int32, Int64}, or CIndex{Int32}, CIndex{Int64}, if the indices are already 0-based. Valid element types are {Float32, Float64, ComplexF64}.

Arguments

(rows, cols), v : sparse storage vectors, must be valid COO matrix internals with types noted above.

firstrow : the 1-based starting row of the matrix on this rank.

localsize : the local size of the matrix on this rank.

globalsize : the size of the matrix across all ranks.

grid : the grid on which the matrix is distributed.

SuperLUDIST.ReplicatedSuperMatrix — Type

mutable struct ReplicatedSuperMatrix{T, I, S, F, J, G} <: SuperLUDIST.AbstractSuperMatrix{T, I, S}

SuperMatrix that has been globally replicated on all ranks in a communicator. The local storage contains the full matrix, the Julia representation may be accessed by A.store::SparseBase.CSCStore{T, CIndex{<:Int}}.

Extended

supermatrix

Internal: Reference to Common.SuperMatrix{I}

format

Internal: Reference to storage format Common.NCformat{I} held by supermatrix

store

Julia-side storage, of type SparseBase.CSCStore. Keeps memory alive.

grid

MatrixMarket.mmread — Method

mmread(DistributedSuperMatrix{Tv, Ti}, filename::String, partitioner = distribute_evenly)

Read an mtx file into a compressed sparse column format and broadcast it to all ranks in comm.

To read an mtx file locally use mmread(SparseBase.CSCStore...).

Arguments

arg1<:Type{ReplicatedSuperMatrix{Tv, Ti}}

Tv <: Union{Float32, Float64, ComplexF64}: Non-optional eltype.

Ti <: Union{Int32, Int64}: Non-optional index type.

filename::String: .mtx file accessible by the root process.

Keyword Arguments

desymmetrize::Bool = true: If the matrix file is represented

in symmetric form, represent in the full form.

root::Int = 0: The MPI process on which to read the file.

comm::MPI.Comm = MPI.COMM_WORLD: The MPI communicator on which to distribute

the matrix.

Example

using MPI, SuperLUDIST, MatrixMarket
A = MatrixMarket.mmread(ReplicatedSuperMatrix{Float64, Int32}, )

MatrixMarket.mmread — Method

mmread(ReplicatedSuperMatrix{Tv, Ti}, filename::String)

Read an mtx file into a compressed sparse column format and broadcast it to all ranks in comm.

To read an mtx file locally use mmread(SparseBase.CSCStore...).

Arguments

arg1<:Type{ReplicatedSuperMatrix{Tv, Ti}}

Tv <: Union{Float32, Float64, ComplexF64}: Non-optional eltype.

Ti <: Union{Int32, Int64}: Non-optional index type.

filename::String: .mtx file accessible by the root process.

Keyword Arguments

desymmetrize::Bool = true: If the matrix file is represented

in symmetric form, represent in the full form.

root::Int = 0: The MPI process on which to read the file.

comm::MPI.Comm = MPI.COMM_WORLD: The MPI communicator on which to distribute

the matrix.

Example

using MPI, SuperLUDIST, MatrixMarket
A = MatrixMarket.mmread(ReplicatedSuperMatrix{Float64, Int32}, )

SparseBase.Communication.scatterstore! — Method

scatterstore!(rstore::DistributedSuperMatrix{Tv, Ti}, sstore::Union{Nothing, SparseBase.SinglyCompressedStore{Tvalues, StorageOrders.RowMajor()} where Tvalues}, chunksizes) -> DistributedSuperMatrix

SuperLUDIST.pgssvx! — Method

pgssvx!(A::DistributedSuperMatrix{Tv, Ti}, b::Union{Array{Tv, 1}, Array{Tv, 2}}) -> Tuple{Any, Any}

Solve the distributed sparse linear system Ax = b, overwriting b with the solution.

Returns b and the factor object which may be used to solve with different b or reuse parts of the factorization different values values for A.

SuperLUDIST.pgssvx! — Method

pgssvx!(A::ReplicatedSuperMatrix{Tv, Ti}, b::Union{Array{Tv, 1}, Array{Tv, 2}}) -> Tuple{Any, Any}

Solve the replicated sparse linear system Ax = b, overwriting b with the solution.

Returns b and the factor object which may be used to solve with different b or reuse parts of the factorization different values values for A.